src/appv2/requirements.txt
dotenv
chainlit
openai
azure-cosmos
azure-storage-blob
azure-ai-projects
azure-identity
asyncpg
psycopg2-binary
agent-framework
mcp[cli]